requirements.txt
google-cloud-bigquery==1.26.1
google-oauth2-tool==0.0.3